use same `string to float convertion logic` in both `CameraControl` and `CutPlaneMenu`
